feat: apprenticeship pass - the education system comes alive (Iteration 32)
- Audit: the education module (§8.1.18 - attempt_teaching/record_learning/effective_teaching) was fully implemented but completely dead - only daily_update was called. Its u32 knowledge ids mismatched the sim u64 knowledge store, proving it was never wired. This was the plan Phase-3 education/apprenticeship priority.

- Fix 1: aligned education knowledge ids u32 -> u64 to match CulturalState.knowledge.

- Fix 2: run_apprenticeship_pass in the daily block - each agent learns one knowledge item they lack from the best capable teacher (teaching_skill gate + relationship quality via new relationship_quality helper). Scans candidates until a teachable item is found (a student missing several items may only have a teacher for one). On success: records education events, transfers to cultural knowledge, bumps holders, fires the Iter-28 desacralization hook (evidence = learning_rate), pushes KnowledgeTransferred event + provenance.

- Tests: 2 wiring tests (transfer succeeds with warm relationship + skilled teacher; no-teacher no-op). Education module tests still pass after type change.

- Validation: full workspace 692 passed (+2), 0 failed; zero snapshot/golden drift (agents already knew most seeded knowledge, transfers are rare and tiny); clippy clean on new code; no debug, no RNG; live 30K identical macro timeline (faction tick 6676, revolutions intact).

+++ crates/mindstrata-sim/src/culture/education.rs
@@ -23,8 +23,11 @@ pub struct EducationEvent {
     pub teacher: usize,
     /// Index of the student agent.
     pub student: usize,
-    /// Knowledge id being transmitted.
-    pub knowledge_id: u32,
+    /// Knowledge id being transmitted. `u64` to match the sim's knowledge
+    /// store (CulturalState.knowledge: Vec<u64>) — the education module was
+    /// originally written with `u32` ids and never wired in; alignment is a
+    /// precondition of the apprenticeship pass.
+    pub knowledge_id: u64,
     /// Quality of the teaching (0–1).
     pub quality: Fixed,
     /// Student's learning rate (depends on aptitude, motivation, relationship).
@@ -39,9 +42,9 @@ pub struct EducationEvent {
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
 pub struct EducationState {
     /// Knowledge ids this agent has learned.
-    pub learned: Vec<u32>,
+    pub learned: Vec<u64>,
     /// Knowledge ids this agent is currently learning (in progress).
-    pub in_progress: Vec<u32>,
+    pub in_progress: Vec<u64>,
     /// Teaching skill — how well this agent can teach others (0–1).
     pub teaching_skill: Fixed,
     /// Learning aptitude — how quickly this agent learns (0–1).
@@ -76,7 +79,7 @@ impl EducationState {
     }
 
     /// Check if this agent has learned a specific knowledge id.
-    pub fn has_learned(&self, knowledge_id: u32) -> bool {
+    pub fn has_learned(&self, knowledge_id: u64) -> bool {
         self.learned.contains(&knowledge_id)
     }
 
@@ -122,7 +125,7 @@ impl EducationState {
 pub fn attempt_teaching(
     teacher_idx: usize,
     student_idx: usize,
-    knowledge_id: u32,
+    knowledge_id: u64,
     teacher_education: &EducationState,
     student_education: &EducationState,
     knowledge_familiarity: Fixed,
